convert economy row to list item
