+++ src/renderer/components/economy/economy.css
@@ -39,14 +39,13 @@
 }
 
 .economy_title {
+  margin: 4px 0;
   display: grid;
-  padding: 0px 16px;
+  padding: 0px 8px;
   grid-template-columns: auto 10% 15% 20% 20% 15%;
   color: var(--color-text);
   line-height: 64px;
-  font-family: "belerenbold" !important;
   font-size: 16px;
-  background-color: var(--color-section-active);
 }
 
 .economy_sub {
