+++ src/renderer/components/list-item/ListItem.css
@@ -5,10 +5,10 @@
   cursor: pointer;
   height: 64px;
   display: flex;
-  margin: 8px 0px;
+  margin: 4px 0px;
   border-radius: 4px;
   justify-content: space-between;
-  background-color: var(--color-section);
+  background-color: var(--color-dark);
   border: 1px solid var(--color-dark);
 }
